requirements.txt
fastapi
aioredis
pymongo[srv]
pydantic
pydantic[email]
uvicorn
python-dotenv
pyjwt
cryptography
